renaming SecComplex to Complex again
